+++ app/src/main/java/woowacourse/shopping/data/mapper/ShoppingCartMapper.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data.mapper
 
-import woowacourse.shopping.data.shoppingcart.ShoppingCartEntity
+import woowacourse.shopping.data.cart.ShoppingCartEntity
 import woowacourse.shopping.domain.ShoppingProduct
 
 fun List<ShoppingCartEntity>.toDomain() =
